requirements.txt
groq
python-dotenv
pyautogui
Pillow
# For OCR (if we implement local text reading)
pytesseract
